requirements.txt
streamlit==1.28.1
firecrawl-py>=0.0.16
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
